Search functions in a package matching a pattern.

You may need a function which name you remember only partly. Use this function to serach all functions which name match whatever you remember.

search_function(pattern, pkg_nm)

Arguments

pattern
Argument passed to base::grepl().

pkg_nm
A character string giving the name of the package to search.

Value

A vector of characters

Details

There are multiple ways to find a function in a package, e.g. with the autocomplete feature in RStudio; by searching the package's index; and by searching the function's reference in the package's website. However, there is no way to do it programmatically. This function provides a programmatic way to search functions in a package, based on a user-defined pattern, and makes it easy to search multiple patterns at once via lapply().

See also

base::grepl()

Examples

# R loads a number of packages automatically. See what those packages are:
search()

#>  [1] ".GlobalEnv"                 "package:bindrcpp"          
#>  [3] "package:DependenciesGraphs" "package:XML"               
#>  [5] "package:shinydashboard"     "package:visNetwork"        
#>  [7] "package:handy"              "devtools_shims"            
#>  [9] "mvb.session.info"           "package:rhub"              
#> [11] "tools:rstudio"              "package:stats"             
#> [13] "package:graphics"           "package:grDevices"         
#> [15] "package:utils"              "package:datasets"          
#> [17] "package:testthat"           "package:usethis"           
#> [19] "package:devtools"           "package:methods"           
#> [21] "Autoloads"                  "package:base"              

# Directly search functions in any package already loaded: e.g. "base"
search_function("summary", pkg_nm = "base")

#>  [1] "format.summaryDefault"   "print.summary.table"    
#>  [3] "print.summaryDefault"    "summary"                
#>  [5] "summary.connection"      "summary.data.frame"     
#>  [7] "Summary.data.frame"      "summary.Date"           
#>  [9] "Summary.Date"            "summary.default"        
#> [11] "Summary.difftime"        "summary.factor"         
#> [13] "Summary.factor"          "summary.matrix"         
#> [15] "Summary.numeric_version" "Summary.ordered"        
#> [17] "summary.POSIXct"         "Summary.POSIXct"        
#> [19] "summary.POSIXlt"         "Summary.POSIXlt"        
#> [21] "summary.proc_time"       "summary.srcfile"        
#> [23] "summary.srcref"          "summary.table"          

# Case is ignored
search_function("Summary", pkg_nm = "base")

#>  [1] "format.summaryDefault"   "print.summary.table"    
#>  [3] "print.summaryDefault"    "summary"                
#>  [5] "summary.connection"      "summary.data.frame"     
#>  [7] "Summary.data.frame"      "summary.Date"           
#>  [9] "Summary.Date"            "summary.default"        
#> [11] "Summary.difftime"        "summary.factor"         
#> [13] "Summary.factor"          "summary.matrix"         
#> [15] "Summary.numeric_version" "Summary.ordered"        
#> [17] "summary.POSIXct"         "Summary.POSIXct"        
#> [19] "summary.POSIXlt"         "Summary.POSIXlt"        
#> [21] "summary.proc_time"       "summary.srcfile"        
#> [23] "summary.srcref"          "summary.table"          

# Easily search multiple patterns with lapply
multiple_patterns <- c("min", "ply")
lapply(multiple_patterns, search_function, pkg_nm = "base")

#> [[1]]
#> [1] "cummin"             "determinant"        "determinant.matrix"
#> [4] "min"                "pmin"               "pmin.int"          
#> [7] "which.min"         
#> 
#> [[2]]
#> [1] "apply"  "eapply" "lapply" "mapply" "rapply" "sapply" "tapply" "vapply"
#> 

# If the package you want to search isn't load it, you must load it first.

# NOT RUN {
library(ggplot2)
search_function("point", pkg_nm = "ggplot2")
# }